+++ packages/testkit-backend/src/channel/index.js
@@ -1,4 +1,4 @@
-import Channel from "./abstract"
+import Channel from "./interface"
 import SocketChannel from "./socket"
 
 export default Channel
