+++ packages/testkit-backend/src/controller/index.js
@@ -1,0 +1,7 @@
+import Controller from './interface'
+import LocalController from './local'
+
+export default Controller
+export {
+  LocalController
+}
