+++ packages/testkit-backend/src/index.js
@@ -1,10 +1,11 @@
 import Backend from './backend'
 import { SocketChannel } from './channel'
-import NodeController from './node.controller'
+import { LocalController } from './controller'
+import * as request_handlers from './request-handlers'
 
 function main( ) {
   const newChannel = () => new SocketChannel(process.env.BACKEND_PORT || 9876)
-  const newController = () => new NodeController()
+  const newController = () => new LocalController(request_handlers)
   const backend = new Backend(newController, newChannel)
     
   backend.start()
